Refactor: Remove 'view' prefix from as many modules as possible...
...without raising errors. viewQuestionnaireSettings needs
some attention :face_exhaling:

Refs: BLAIS5-3651
