+++ src/components/questionnaireDetailsPage/catiModeDetails.test.tsx
@@ -6,7 +6,7 @@ import flushPromises from "../../tests/utils";
 import { render, waitFor, screen } from "@testing-library/react";
 import { act } from "react-dom/test-utils";
 import React from "react";
-import ViewCatiModeDetails from "./viewCatiModeDetails";
+import CatiModeDetails from "./catiModeDetails";
 import axios from "axios";
 import MockAdapter from "axios-mock-adapter";
 
@@ -19,7 +19,7 @@ describe("View TO Start Date section", () => {
 
     it("should not render for non-CATI questionnaires", async () => {
         const { container } = render(
-            <ViewCatiModeDetails questionnaireName={"OPN2101A"} modes={["CAWI"]} />
+            <CatiModeDetails questionnaireName={"OPN2101A"} modes={["CAWI"]} />
         );
 
         await act(async () => {
@@ -35,7 +35,7 @@ describe("View TO Start Date section", () => {
         const viewToStartDateFailedMessage = /Failed to get Telephone Operations start date/i;
         mock.onGet("/api/tostartdate/OPN2101A").reply(500);
         render(
-            <ViewCatiModeDetails questionnaireName={"OPN2101A"} modes={["CATI"]} />
+            <CatiModeDetails questionnaireName={"OPN2101A"} modes={["CATI"]} />
         );
 
         await act(async () => {
