+++ src/components/questionnaireDetailsPage/catiModeDetails.tsx
@@ -1,39 +1,40 @@
 import React, { ReactElement, useEffect, useState } from "react";
 import { ONSLoadingPanel, ONSPanel } from "blaise-design-system-react-components";
-import { getTMReleaseDate } from "../../client/tmReleaseDate";
+import { getTOStartDate } from "../../client/toStartDate";
 import dateFormatter from "dayjs";
 import TimeAgo from "react-timeago";
 import { Link } from "react-router-dom";
 
 interface Props {
     questionnaireName: string
+    modes: string[]
 }
 
-function ViewTmDetails({ questionnaireName }: Props): ReactElement {
-    if (!questionnaireName.includes("LMS")) {
+function CatiModeDetails({ questionnaireName, modes }: Props): ReactElement {
+    if (!modes.includes("CATI")) {
         return <></>;
     }
 
     const [loading, setLoading] = useState<boolean>(true);
     const [errored, setErrored] = useState<boolean>(false);
-    const [tmReleaseDate, setTmReleaseDate] = useState<boolean>(false);
-    const [tmReleaseDateValue, setTmReleaseDateValue] = useState<string>("");
+    const [toStartDate, setToStartDate] = useState<boolean>(false);
+    const [toStartDateValue, setToStartDateValue] = useState<string>("");
 
     useEffect(() => {
-        getTMReleaseDateForQuestionnaire().then(() => setLoading(false));
+        getTOStartDateForQuestionnaire().then(() => setLoading(false));
     }, []);
 
-    async function getTMReleaseDateForQuestionnaire() {
+    async function getTOStartDateForQuestionnaire() {
         setLoading(true);
         try {
-            const tmReleaseDate = await getTMReleaseDate(questionnaireName);
-            if (tmReleaseDate == "") {
-                setTmReleaseDate(false);
+            const toStartDate = await getTOStartDate(questionnaireName);
+            if (toStartDate == "") {
+                setToStartDate(false);
                 return;
             }
 
-            setTmReleaseDate(true);
-            setTmReleaseDateValue(tmReleaseDate);
+            setToStartDate(true);
+            setToStartDateValue(toStartDate);
         } catch {
             setErrored(true);
         }
@@ -42,7 +43,7 @@ function ViewTmDetails({ questionnaireName }: Props): ReactElement {
     if (loading) {
         return (
             <div className="ons-u-mb-m" aria-busy="true">
-                <ONSLoadingPanel message={"Getting Totalmobile release date"}/>
+                <ONSLoadingPanel message={"Getting Telephone Operations start date"}/>
             </div>
         );
     }
@@ -50,7 +51,7 @@ function ViewTmDetails({ questionnaireName }: Props): ReactElement {
     if (errored) {
         return (
             <div className="ons-u-mb-m">
-                <ONSPanel status={"error"}>Failed to get Totalmobile release date</ONSPanel>
+                <ONSPanel status={"error"}>Failed to get Telephone Operations start date</ONSPanel>
             </div>
         );
     }
@@ -59,7 +60,7 @@ function ViewTmDetails({ questionnaireName }: Props): ReactElement {
         <>
             <div className="ons-summary ons-u-mb-m elementToFadeIn">
                 <div className="ons-summary__group">
-                    <h2 className="ons-summary__group-title">Totalmobile details</h2>
+                    <h2 className="ons-summary__group-title">CATI mode details</h2>
                     <table className="ons-summary__items">
                         <thead className="ons-u-vh">
                             <tr>
@@ -72,39 +73,39 @@ function ViewTmDetails({ questionnaireName }: Props): ReactElement {
                             <tr className="ons-summary__row ons-summary__row--has-values">
                                 <td className="ons-summary__item-title">
                                     <div className="ons-summary__item--text">
-                                    Totalmobile release date
+                                    Telephone Operations start date
                                     </div>
                                 </td>
                                 <td className="ons-summary__values">
 
                                     {
-                                        tmReleaseDate ?
+                                        toStartDate ?
                                             <>
-                                                {dateFormatter(tmReleaseDateValue).format("DD/MM/YYYY")} ({<TimeAgo
-                                                    live={false} date={tmReleaseDateValue}/>})
+                                                {dateFormatter(toStartDateValue).format("DD/MM/YYYY")} ({<TimeAgo
+                                                    live={false} date={toStartDateValue}/>})
                                             </>
 
                                             :
-                                            "No release date specified"
+                                            "No start date specified, using survey days"
                                     }
                                 </td>
                                 <td className="ons-summary__actions">
                                     {
-                                        tmReleaseDate ?
+                                        toStartDate ?
                                             <Link to={{
-                                                pathname: "/questionnaire/release-date",
-                                                state: { questionnaireName: questionnaireName, tmReleaseDate: tmReleaseDateValue }
+                                                pathname: "/questionnaire/start-date",
+                                                state: { questionnaireName: questionnaireName, toStartDate: toStartDateValue }
                                             }} className="ons-summary__button"
-                                            aria-label={`Change or delete release date for questionnaire ${questionnaireName}`}>
-                                            Change or delete release date
+                                            aria-label={`Change or delete start date for questionnaire ${questionnaireName}`}>
+                                            Change or delete start date
                                             </Link>
                                             :
                                             <Link to={{
-                                                pathname: "/questionnaire/release-date",
+                                                pathname: "/questionnaire/start-date",
                                                 state: { questionnaireName: questionnaireName }
                                             }} className="ons-summary__button"
-                                            aria-label={`Add a release date for questionnaire ${questionnaireName}`}>
-                                            Add release date
+                                            aria-label={`Add a start date for questionnaire ${questionnaireName}`}>
+                                            Add start date
                                             </Link>
 
                                     }
@@ -118,4 +119,4 @@ function ViewTmDetails({ questionnaireName }: Props): ReactElement {
     );
 }
 
-export default ViewTmDetails;
+export default CatiModeDetails;
