+++ src/components/questionnaireDetailsPage/totalmobileDetails.test.tsx
@@ -6,7 +6,7 @@ import flushPromises from "../../tests/utils";
 import { render, screen } from "@testing-library/react";
 import { act } from "react-dom/test-utils";
 import React from "react";
-import ViewTmDetails from "./viewTmDetails";
+import TotalmobileDetails from "./totalmobileDetails";
 import axios from "axios";
 import MockAdapter from "axios-mock-adapter";
 import { Router } from "react-router-dom";
@@ -25,7 +25,7 @@ describe("View Totalmobile details", () => {
         mock.onGet("/api/tmreleasedate/LMS2101_AA1").reply(200, { tmreleasedate: "2021-06-27T16:29:00+00:00" });
         const rerender = render(
             <Router history={history}>
-                <ViewTmDetails questionnaireName={"LMS2101_AA1"}/>
+                <TotalmobileDetails questionnaireName={"LMS2101_AA1"}/>
             </Router>
         );
 
@@ -45,7 +45,7 @@ describe("View Totalmobile details", () => {
         mock.onGet("/api/tmreleasedate/LMS2101_AA1").reply(404, { tmreleasedate: "" });
         const rerender = render(
             <Router history={history}>
-                <ViewTmDetails questionnaireName={"LMS2101_AA1"}/>
+                <TotalmobileDetails questionnaireName={"LMS2101_AA1"}/>
             </Router>
         );
 
@@ -64,7 +64,7 @@ describe("View Totalmobile details", () => {
         mock.onGet("/api/tmreleasedate/LMS2101A").reply(500);
         const rerender = render(
             <Router history={history}>
-                <ViewTmDetails questionnaireName={"LMS2101A"}/>
+                <TotalmobileDetails questionnaireName={"LMS2101A"}/>
             </Router>
         );
 
@@ -81,7 +81,7 @@ describe("View Totalmobile details", () => {
         mock.onGet("/api/tmreleasedate/OPN2101_AA1").reply(200, { tmreleasedate: "2021-06-27T16:29:00+00:00" });
         const rerender = render(
             <Router history={history}>
-                <ViewTmDetails questionnaireName={"OPN2101_AA1"}/>
+                <TotalmobileDetails questionnaireName={"OPN2101_AA1"}/>
             </Router>
         );
