+++ src/components/questionnaireDetailsPage/totalmobileDetails.tsx
@@ -1,40 +1,39 @@
 import React, { ReactElement, useEffect, useState } from "react";
 import { ONSLoadingPanel, ONSPanel } from "blaise-design-system-react-components";
-import { getTOStartDate } from "../../client/toStartDate";
+import { getTMReleaseDate } from "../../client/tmReleaseDate";
 import dateFormatter from "dayjs";
 import TimeAgo from "react-timeago";
 import { Link } from "react-router-dom";
 
 interface Props {
     questionnaireName: string
-    modes: string[]
 }
 
-function ViewCatiModeDetails({ questionnaireName, modes }: Props): ReactElement {
-    if (!modes.includes("CATI")) {
+function TotalmobileDetails({ questionnaireName }: Props): ReactElement {
+    if (!questionnaireName.includes("LMS")) {
         return <></>;
     }
 
     const [loading, setLoading] = useState<boolean>(true);
     const [errored, setErrored] = useState<boolean>(false);
-    const [toStartDate, setToStartDate] = useState<boolean>(false);
-    const [toStartDateValue, setToStartDateValue] = useState<string>("");
+    const [tmReleaseDate, setTmReleaseDate] = useState<boolean>(false);
+    const [tmReleaseDateValue, setTmReleaseDateValue] = useState<string>("");
 
     useEffect(() => {
-        getTOStartDateForQuestionnaire().then(() => setLoading(false));
+        getTMReleaseDateForQuestionnaire().then(() => setLoading(false));
     }, []);
 
-    async function getTOStartDateForQuestionnaire() {
+    async function getTMReleaseDateForQuestionnaire() {
         setLoading(true);
         try {
-            const toStartDate = await getTOStartDate(questionnaireName);
-            if (toStartDate == "") {
-                setToStartDate(false);
+            const tmReleaseDate = await getTMReleaseDate(questionnaireName);
+            if (tmReleaseDate == "") {
+                setTmReleaseDate(false);
                 return;
             }
 
-            setToStartDate(true);
-            setToStartDateValue(toStartDate);
+            setTmReleaseDate(true);
+            setTmReleaseDateValue(tmReleaseDate);
         } catch {
             setErrored(true);
         }
@@ -43,7 +42,7 @@ function ViewCatiModeDetails({ questionnaireName, modes }: Props): ReactElement 
     if (loading) {
         return (
             <div className="ons-u-mb-m" aria-busy="true">
-                <ONSLoadingPanel message={"Getting Telephone Operations start date"}/>
+                <ONSLoadingPanel message={"Getting Totalmobile release date"}/>
             </div>
         );
     }
@@ -51,7 +50,7 @@ function ViewCatiModeDetails({ questionnaireName, modes }: Props): ReactElement 
     if (errored) {
         return (
             <div className="ons-u-mb-m">
-                <ONSPanel status={"error"}>Failed to get Telephone Operations start date</ONSPanel>
+                <ONSPanel status={"error"}>Failed to get Totalmobile release date</ONSPanel>
             </div>
         );
     }
@@ -60,7 +59,7 @@ function ViewCatiModeDetails({ questionnaireName, modes }: Props): ReactElement 
         <>
             <div className="ons-summary ons-u-mb-m elementToFadeIn">
                 <div className="ons-summary__group">
-                    <h2 className="ons-summary__group-title">CATI mode details</h2>
+                    <h2 className="ons-summary__group-title">Totalmobile details</h2>
                     <table className="ons-summary__items">
                         <thead className="ons-u-vh">
                             <tr>
@@ -73,39 +72,39 @@ function ViewCatiModeDetails({ questionnaireName, modes }: Props): ReactElement 
                             <tr className="ons-summary__row ons-summary__row--has-values">
                                 <td className="ons-summary__item-title">
                                     <div className="ons-summary__item--text">
-                                    Telephone Operations start date
+                                    Totalmobile release date
                                     </div>
                                 </td>
                                 <td className="ons-summary__values">
 
                                     {
-                                        toStartDate ?
+                                        tmReleaseDate ?
                                             <>
-                                                {dateFormatter(toStartDateValue).format("DD/MM/YYYY")} ({<TimeAgo
-                                                    live={false} date={toStartDateValue}/>})
+                                                {dateFormatter(tmReleaseDateValue).format("DD/MM/YYYY")} ({<TimeAgo
+                                                    live={false} date={tmReleaseDateValue}/>})
                                             </>
 
                                             :
-                                            "No start date specified, using survey days"
+                                            "No release date specified"
                                     }
                                 </td>
                                 <td className="ons-summary__actions">
                                     {
-                                        toStartDate ?
+                                        tmReleaseDate ?
                                             <Link to={{
-                                                pathname: "/questionnaire/start-date",
-                                                state: { questionnaireName: questionnaireName, toStartDate: toStartDateValue }
+                                                pathname: "/questionnaire/release-date",
+                                                state: { questionnaireName: questionnaireName, tmReleaseDate: tmReleaseDateValue }
                                             }} className="ons-summary__button"
-                                            aria-label={`Change or delete start date for questionnaire ${questionnaireName}`}>
-                                            Change or delete start date
+                                            aria-label={`Change or delete release date for questionnaire ${questionnaireName}`}>
+                                            Change or delete release date
                                             </Link>
                                             :
                                             <Link to={{
-                                                pathname: "/questionnaire/start-date",
+                                                pathname: "/questionnaire/release-date",
                                                 state: { questionnaireName: questionnaireName }
                                             }} className="ons-summary__button"
-                                            aria-label={`Add a start date for questionnaire ${questionnaireName}`}>
-                                            Add start date
+                                            aria-label={`Add a release date for questionnaire ${questionnaireName}`}>
+                                            Add release date
                                             </Link>
 
                                     }
@@ -119,4 +118,4 @@ function ViewCatiModeDetails({ questionnaireName, modes }: Props): ReactElement 
     );
 }
 
-export default ViewCatiModeDetails;
+export default TotalmobileDetails;
